bridge jour 8

+++ pointers_arrays_strings/100-atoi.c
@@ -13,9 +13,9 @@
 * Return: long int.
 */
 
-long int _strlen(char *s)
+int _strlen(char *s)
 {
-long int len = 0;
+int len = 0;
 
 while (s[len] != '\0')
 {
